+++ client/src/components/Pages.jsx
@@ -5,12 +5,14 @@ import { Pagination } from "react-bootstrap";
 
 const Pages = observer(() => {
   const { device } = useContext(Context);
+  // TODO: у тебя чтото похожее гдето уже писалось, и лучше бы тогда вынести в какую-то функцию или хук
   const pageCount = Math.ceil(device.totalCount / device.limit);
   const pages = [];
 
   for (let i = 1; i <= pageCount; i++) {
     pages.push(i);
   }
+  // вот до сюда
 
   return (
     <div className="pagination">
