+++ client/src/styles/Admin.css
@@ -196,6 +196,7 @@
 }
 
 @media screen and (max-width: 1024px) {
+  /*TODO: избыточный селектор, тк он не меняет вышеописанные стили*/
   .admin-roll-img {
     max-width: 100%;
     max-height: 100%;
@@ -226,3 +227,4 @@
     margin-right: 0vw;
   }
 }
+/*TODO: при нулевых значениях единицы измерения можно не писать */
