+++ client/src/styles/AuthStyle.css
@@ -8,6 +8,10 @@
 
 @import url("https://fonts.googleapis.com/css2?family=Kanit:ital,wght@1,300&display=swap");
 
+/*TODO: куча одинаковых стилей, тк ты все равно все подключаешь глобально (посмотри в девтулзах стили), то стоило бы
+    оставить один экземпляр, тк это увеличивает конечный бандл, или вынести все одинаковые стили, которые ТОЧНО должны быть
+    глобальными в отдельный файл и подключать в index.js или подобном месте
+    */
 * {
   margin: 0;
   padding: 0;
