+++ client/src/styles/Basket.css
@@ -8,6 +8,7 @@
     display: flex;
     /* flex-direction: column; */
 }
+/*TODO: display: grid тоже самое что и flex column*/
 .basket__list-devices{
     display: flex;
     flex-direction: column;
